+++ packages/hedera-identify-snap/packages/snap/src/custom-ui/onHome.ts
@@ -30,15 +30,12 @@ import {
   text,
 } from '@metamask/snaps-sdk';
 import _ from 'lodash';
-import { GetAccountInfoFacade } from '../facades/snap/GetAccountInfoFacade';
+import { getNetworkNameFromChainId } from '../constants';
 import { SnapAccounts } from '../snap/SnapAccounts';
 import { SnapState } from '../snap/SnapState';
-import type { PublicAccountInfo } from '../types/account';
-import type { IdentifySnapParams } from '../types/state';
 import { EvmUtils } from '../utils/EvmUtils';
 
 export const OnHomePageUI: OnHomePageHandler = async () => {
-  // Set origin to be the current page
   const origin = 'Identify Snap';
 
   let state = await SnapState.getStateUnchecked();
@@ -46,10 +43,7 @@ export const OnHomePageUI: OnHomePageHandler = async () => {
     state = await SnapState.initState();
   }
 
-  // Get network
   const network = await EvmUtils.getChainId();
-
-  // Get connected metamask account
   const connectedAddress = await SnapAccounts.getCurrentMetamaskAccount();
 
   const snapAddress = await SnapAccounts.setCurrentAccount(
@@ -61,93 +55,112 @@ export const OnHomePageUI: OnHomePageHandler = async () => {
     true,
   );
 
-  const identifySnapParams: IdentifySnapParams = {
-    origin,
-    state,
+  return {
+    content: panel([
+      heading('Identify Snap'),
+      text('Welcome to Identify Snap home page!'),
+      divider(),
+      text(`**Network**: ${getNetworkNameFromChainId(network)}`),
+      divider(),
+      text(`**Snap Account DID**:`),
+      copyable(state.currentAccount.identifier.did || 'Unavailable'),
+      text(`**Snap Account EVM Address**:`),
+      copyable(snapAddress || 'Unavailable'),
+      text(`**Snap Hedera Account ID**:`),
+      copyable(state.currentAccount.hederaAccountId || 'Unavailable'),
+      text(`**Associated Metamask Address**:`),
+      copyable(connectedAddress || 'Unavailable'),
+      text(`**Current DID Method**: ${state.currentAccount.method || 'None'}`),
+      divider(),
+      text(`**Switch DID Method**:`),
+      form({
+        name: 'form-switch-did-method',
+        children: [
+          input({
+            name: 'didMethod',
+            placeholder: 'Methods: [did:pkh, did:key, did:hedera]',
+            value: state.currentAccount.method || '',
+          }),
+          button({
+            value: 'Switch DID Method',
+            buttonType: 'submit',
+          }),
+        ],
+      }),
+      divider(),
+      text(`**Resolve DID**:`),
+      form({
+        name: 'form-resolve-did',
+        children: [
+          input({
+            name: 'did',
+            placeholder: 'Enter DID string',
+            value: state.currentAccount.identifier.did || '',
+          }),
+          button({
+            value: 'Resolve DID',
+            buttonType: 'submit',
+          }),
+        ],
+      }),
+      divider(),
+      text(`**Create a new Verifiable Credential**:`),
+      form({
+        name: 'form-create-vc',
+        children: [
+          input({
+            name: 'vcText',
+            placeholder: 'Enter some text',
+            value: 'Sample text',
+          }),
+          button({
+            value: 'Create VC',
+            buttonType: 'submit',
+          }),
+        ],
+      }),
+      divider(),
+      text(`**Retrieve Verifiable Credentials**:`),
+      form({
+        name: 'form-get-vcs',
+        children: [
+          button({
+            value: 'Get VCs',
+            buttonType: 'submit',
+          }),
+        ],
+      }),
+      divider(),
+      text(`**Create Verifiable Presentation**:`),
+      form({
+        name: 'form-create-vp',
+        children: [
+          input({
+            name: 'vcData',
+            placeholder: 'Enter your Verifiable Credential',
+            value: '',
+          }),
+          button({
+            value: 'Create VP',
+            buttonType: 'submit',
+          }),
+        ],
+      }),
+      divider(),
+      button({
+        value: 'Export Snap Account Private Key',
+        name: 'btn-export-snap-account-private-key',
+      }),
+      divider(),
+      text(`**FAQs**:`),
+      text(
+        `**Why is my MetaMask address different from my Snap EVM address?**: Since MetaMask does not allow Snaps to access the private keys of MetaMask accounts, Identify Snap creates a new account with its own private key. However, this new account is still associated with the MetaMask account so you can easily get back to your account on any device as long as you're connected to the same MetaMask account.`,
+      ),
+      text(
+        `**Is the Snap safe to use?**: Yes, the Snap is safe to use as it is run in a sandboxed environment and uses a permissions model to protect your data and respect your consent. The Snap does not have access to your MetaMask account data.`,
+      ),
+      divider(),
+      text(`[View Snap documentation](https://docs.tuum.tech/identify)`),
+    ]),
   };
-
-  try {
-    const accountInfo: PublicAccountInfo =
-      await GetAccountInfoFacade.getAccountInfo(identifySnapParams);
-
-    return {
-      content: panel([
-        heading('Identify Snap'),
-        text('Welcome to Identify Snap home page!'),
-        divider(),
-        text(`**Network**: ${network}`),
-        divider(),
-        text(`**Account EVM Address**:`),
-        copyable(accountInfo.snapAddress),
-        text(`**Associated Metamask Address**:`),
-        copyable(connectedAddress),
-        divider(),
-        text(`**Transfer Hbar**:`),
-        form({
-          name: 'form-transfer-crypto',
-          children: [
-            input({
-              name: 'to',
-              placeholder: 'Recipient Address',
-            }),
-            input({
-              name: 'amount',
-              placeholder: 'Amount',
-            }),
-            input({
-              name: 'memo',
-              placeholder: 'Memo(Needed for exchange transfers)',
-            }),
-            button({
-              value: 'Send Hbar',
-              buttonType: 'submit',
-            }),
-          ],
-        }),
-        divider(),
-        button({
-          value: 'Export Snap Account Private Key',
-          name: 'btn-export-snap-account-private-key',
-        }),
-        divider(),
-        text(`**FAQs**:`),
-        text(
-          `**Why is my MetaMask address different from my Hedera EVM address?**: Since MetaMask does not allow Snaps to access the private keys of MetaMask accounts, Hedera Wallet Snap creates a new account with its own private key. However, this new account is still associated with the MetaMask account so you can easily get back to your account on any device as long as you're connected to the same MetaMask account.`,
-        ),
-        text(
-          `**Is the Snap safe to use?**: Yes, the Snap is safe to use as it is run in a sandboxed environment and uses a permissions model to protect your data and respect your consent. The Snap does not have access to your MetaMask account data.`,
-        ),
-        divider(),
-        text(
-          `[View Snap documentation](https://docs.tuum.tech/hedera-wallet-snap)`,
-        ),
-        text(`[Visit Pulse for additional features](https://pulse.tuum.tech)`),
-      ]),
-    };
-  } catch (error) {
-    return {
-      content: panel([
-        heading('Identify Snap'),
-        text('Welcome to Identifyt Snap home page!'),
-        divider(),
-        text(`**Network**: ${network}`),
-        divider(),
-        text(`**Account EVM Address**:`),
-        copyable(snapAddress),
-        text(`**Associated Metamask Address**:`),
-        copyable(connectedAddress),
-        divider(),
-        button({
-          value: 'Export Snap Account Private Key',
-          name: 'btn-export-snap-account-private-key',
-        }),
-        divider(),
-        text(`**FAQs**:`),
-        text(
-          `**Why is my MetaMask address different from my Identify EVM address?**: Since MetaMask does not allow Snaps to access the private keys of MetaMask accounts, Hedera Wallet Snap creates a new account with its own private key. However, this new account is still associated with the MetaMask account so you can easily get back to your account on any device as long as you're connected to the same MetaMask account.`,
-        ),
-        text(`[View Snap documentation](https://docs.tuum.tech/identify-snap)`),
-      ]),
-    };
-  }
 };