+++ packages/hedera-identify-snap/packages/snap/CHANGELOG.md
@@ -1,11 +1,18 @@
 # Changelog
 
-## [0.2.0](https://github.com/hashgraph/hedera-metamask-snaps/compare/v0.2.0...v0.2.0) (2024-10-11)
+## 0.2.0 (2024-12-12)
 
 ### :page_with_curl: Documentation
 
 - Refer to the [Identify Snap Wiki](https://docs.tuum.tech/identify) for more info on how the snap works and how to integrate it into your own application.
 
 ### :rocket: Features
+
+- Added support for did:key method
+- Added support for did:hedera method
+- Added snap home page
+- Updated some dialog boxes
+- Upgraded dependencies and optimized the code
+- Refactored the entire codebase
 
 ### :bug: Bug Fixes